refactor: 테스트코드 통과위해 useCart훅 되살림
